+++ src/index.ts
@@ -1,4 +1,5 @@
-import * as types from "./models";
+import * as VnstockTypes from "./models";
+import * as VnstockRealtime from "./core/realtime";
 import { Vnstock } from "./runtime";
 import { createStockAPI, createCommodityAPI } from "./simple";
 
@@ -7,4 +8,4 @@ const stock = createStockAPI(vnstock);
 const commodity = createCommodityAPI(vnstock);
 
 export default vnstock;
-export { types, Vnstock, stock, commodity };
+export { VnstockTypes, Vnstock, VnstockRealtime, stock, commodity };