+++ src/simple.ts
@@ -1,24 +1,32 @@
+import realtime from "./core/realtime";
+import Company from "./core/stock/company";
 import { Vnstock } from "./runtime";
 import { INDEX_SYMBOLS } from "./shared/constants";
-import Company from "./core/stock/company";
 
 export const createStockAPI = (vnstock: Vnstock) => ({
-  price: (ticker: string, start: string, end?: string) => vnstock.stock.quote.history({ symbols: [ticker], start, end, timeFrame: "1D" }),
-  quote: (ticker: string, start: string, end?: string) => vnstock.stock.quote.history({ symbols: [ticker], start, end, timeFrame: "1D" }),
+  quote: ({ ticker, start, end }: { ticker: string; start: string; end?: string }) =>
+    vnstock.stock.quote.history({ symbols: [ticker], start, end, timeFrame: "1D" }),
 
-  index: (index: string, start: string, end?: string) => {
+  priceBoard: ({ ticker }: { ticker: string }) => vnstock.stock.trading.priceBoard([ticker]),
+  topGainers: () => vnstock.stock.trading.topGainers({}),
+  topLosers: () => vnstock.stock.trading.topLosers({}),
+
+  index: ({ index, start, end }: { index: string; start: string; end?: string }) => {
     if (!INDEX_SYMBOLS.includes(index)) {
       throw new Error(`Invalid index symbol: ${index}`);
     }
     return vnstock.stock.quote.history({ symbols: [index], start, end, timeFrame: "1D" });
   },
 
-  company: (ticker: string) => {
+  company: ({ ticker }: { ticker: string }) => {
     vnstock.stock.company = new Company(ticker);
     return vnstock.stock.company.overview();
   },
 
-  financials: (ticker: string, period: "quarter" | "year" = "quarter") => vnstock.stock.financials.balanceSheet({ symbol: ticker, period }),
+  financials: ({ ticker, period = "quarter" }: { ticker: string; period?: "quarter" | "year" }) =>
+    vnstock.stock.financials.balanceSheet({ symbol: ticker, period }),
+
+  realtime,
 });
 
 export const createCommodityAPI = (vnstock: Vnstock) => ({